support reactions for TextMessages (in+out) and VoiceMessages (in+out)
