+++ app/src/main/res/layout/item_custom_incoming_text_message.xml
@@ -86,23 +86,9 @@
             android:textIsSelectable="false"
             app:layout_alignSelf="center" />
 
-
-        <LinearLayout
-            android:id="@+id/reactions_emoji_wrapper"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            app:layout_wrapBefore="true"
-            android:orientation="horizontal"
-            android:layout_marginTop="5dp">
-            <TextView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                tools:text="emojis">
-            </TextView>
-        </LinearLayout>
+        <include
+            android:id="@+id/reactions"
+            layout="@layout/reactions_inside_message" />
 
     </com.google.android.flexbox.FlexboxLayout>
-
-
-
 </RelativeLayout>
