+++ app/src/main/res/layout/item_custom_incoming_voice_message.xml
@@ -108,5 +108,9 @@
                     app:layout_alignSelf="center"
                     tools:text="12:38"/>
 
+            <include
+                android:id="@+id/reactions"
+                layout="@layout/reactions_inside_message" />
+
         </com.google.android.flexbox.FlexboxLayout>
 </RelativeLayout>
